requirements.txt
# Core dependencies
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Document processing
langchain>=0.0.267
langchain-community>=0.0.10
unstructured>=0.10.0
pdfminer.six>=20221105
markdown>=3.4.3
beautifulsoup4>=4.12.2
python-docx>=0.8.11

# Embedding
sentence-transformers>=2.2.2

# Vector database
qdrant-client>=1.4.0

# Testing
pytest>=7.3.1
pytest-cov>=4.1.0

# Utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0 